requirements.txt
cat > requirements.txt << 'EOF'
riscv-reg-block
pytest-cov
streamlit
plotly
networkx
pylint
EOF